Migrate transformRegistrationRequest  logic to registerHelper and streamline registration workflows.
