+++ packages/core-sdk/test/integration/ipAsset.test.ts
@@ -28,7 +28,7 @@ const expect = chai.expect;
 const pool = evenSplitGroupPoolAddress[aeneid];
 const walletAddress = process.env.TEST_WALLET_ADDRESS! as Address;
 
-describe("IP Asset Functions", () => {
+describe.only("IP Asset Functions", () => {
   let client: StoryClient;
   let noCommercialLicenseTermsId: bigint;
   let parentIpId: Hex;
